pyproject.toml
[project]
name = "tina4-python"
version = "0.2.98"
description = "Tina4Python - This is not another framework for Python"
authors = [
    {name = "Andre van Zuydam",email = "andrevanzuydam@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "jinja2>=3.1.5,<4.0.0",
    "libsass (>=0.23.0,<0.24.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "cryptography (>=44.0.0,<45.0.0)",
    "watchdog (>=6.0.0,<7.0.0)",
    "bcrypt (>=4.2.1,<5.0.0)",
    "litequeue (>=0.9,<0.10)",
    "simple-websocket (>=1.1.0,<2.0.0)",
    "hypercorn (>=0.17.3,<0.18.0)",
    "asyncer>=0.0.8",
    "pika>=1.3.2",
]

[dependency-groups]
dev = [
    "flake8>=7.2.0",
    "jurigged>=0.6.0",
    "mysql-connector-python>=9.3.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.5",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "ruff>=0.11.9",
    "safety>=3.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
    "tina4_python/**/*"
]


